requirements.txt
boto3>=1.10.28,<2
botocore
PyYAML
coverage>=5.5
pytest>=6.2.2
pytest-cov>=2.11.1
cfn-lint>=0.45.0
j2cli>=0.3.10
